Add user schema
We have a basic user schema and this ensures that we do the minimum to
begin work on the app.

I opted for a GraphQL setup as it is a great tool for front-end
development, allowing the view layer to define the data it needs rather
than spreading this across the back-end and the front-end.