Add React app to display users
We want a React app that can load user data from the server and then
display it in a table.

I opted to use Apollo to streamline the access to the GraphQL server.

Specifically apollo-boost gives us some conventions that work out of the
box.

Notice that we had to customise the client to pass through the CRSF
token with the request to ensure that the Rails app would respond
correctly and not give a 422 message.